% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0018_applicability.R
\name{check_applicability_power_only_acm0018}
\alias{check_applicability_power_only_acm0018}
\title{Check ACM0018 power-only plant configuration}
\usage{
check_applicability_power_only_acm0018(exports_heat, generates_electricity)
}
\arguments{
\item{exports_heat}{Logical vector indicating whether the facility exports
heat or steam.}

\item{generates_electricity}{Logical vector confirming that the project
generates electricity for delivery to the grid.}
}
\value{
Logical vector confirming that each facility satisfies the power-only
configuration criteria.
}
\description{
Ensures that the project activity is a power-only plant and that no steam or
heat is exported to external users.
}
\examples{
check_applicability_power_only_acm0018(c(FALSE, TRUE), c(TRUE, TRUE))
}
