% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0018_ontology.R
\name{read_acm0018_grid_shapes}
\alias{read_acm0018_grid_shapes}
\title{Load ACM0018 grid connection applicability shape}
\usage{
read_acm0018_grid_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shaclR:read_shacl]{shaclR::read_shacl()}}.
}
\description{
Load ACM0018 grid connection applicability shape
}
